Add errors.{hpp,cpp} translation unit
