+++ include/codeowners/codeowners.hpp
@@ -10,22 +10,6 @@ namespace co
 
 struct repository_impl;
 
-class error : public std::exception
-{
-public:
-    error(const std::string& message)
-        : m_message { message } {};
-    const char* what() const noexcept override { return m_message.c_str(); }
-
-private:
-    std::string m_message;
-};
-
-class repository_not_found_error : public error
-{
-public:
-    using error::error;
-};
 
 class repository
 {
