+++ src/errors.cpp
@@ -1,0 +1,1 @@
+#include <codeowners/errors.hpp>
